Refactoring Navbar to make more managable from Material UI template. In Progress
