--- frontend/src/app/userDashboard/navbar/index.js
@@ -1,1 +1,0 @@
-export * from "./Navbar";
